feat: experimental settings
