+++ lib/app/home/_widgets/eew_card.dart
