+++ lib/app/home/_widgets/location_button.dart
@@ -2,8 +2,8 @@
 library;
 
 import 'package:dpip/api/model/location/location.dart';
-import 'package:dpip/app/home_old/_models/home_location.dart';
-import 'package:dpip/app/home_old/_widgets/blurred_button.dart';
+import 'package:dpip/app/home/_models/home_location.dart';
+import 'package:dpip/app/home/_widgets/blurred_button.dart';
 import 'package:dpip/core/i18n.dart';
 import 'package:dpip/global.dart';
 import 'package:dpip/models/settings/location.dart';
