+++ lib/app/new_home/_widgets/all_observation_average.dart
@@ -1,7 +1,7 @@
 /// A card widget showing the nearest-station temperature and humidity summary.
 library;
 
-import 'package:dpip/app/home/_models/home_model.dart';
+import 'package:dpip/app/new_home/_models/home_model.dart';
 import 'package:dpip/utils/extensions/build_context.dart';
 import 'package:dpip/widgets/typography.dart';
 import 'package:flutter/material.dart';
