+++ lib/app/new_home/_widgets/assistant_hint.dart
@@ -1,7 +1,7 @@
 /// A card widget that shows a contextual weather hint message.
 library;
 
-import 'package:dpip/app/home/_models/home_model.dart';
+import 'package:dpip/app/new_home/_models/home_model.dart';
 import 'package:dpip/utils/extensions/build_context.dart';
 import 'package:flutter/material.dart';
 import 'package:material_symbols_icons/material_symbols_icons.dart';
