+++ lib/app/new_home/_widgets/location_chip.dart
@@ -3,7 +3,7 @@ library;
 
 import 'package:collection/collection.dart';
 import 'package:dpip/api/model/location/location.dart';
-import 'package:dpip/app/home/_models/home_model.dart';
+import 'package:dpip/app/new_home/_models/home_model.dart';
 import 'package:dpip/core/i18n.dart';
 import 'package:dpip/global.dart';
 import 'package:dpip/models/settings/location.dart';
