+++ lib/app/settings/theme/page.dart
@@ -23,64 +23,60 @@ class SettingsThemePage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Consumer<SettingsUserInterfaceModel>(
-      builder: (context, model, child) {
-        return ListView(
+    return ListView(
+      children: [
+        SettingsHeader(
+          icon: Symbols.palette_rounded,
+          title: Text('主題'.i18n),
+          subtitle: Text('調整 DPIP 整體的外觀與顏色'.i18n),
+        ),
+        const SizedBox(height: 16),
+        SegmentedList(
           children: [
-            SettingsHeader(
-              icon: Symbols.palette_rounded,
-              title: Text('主題'.i18n),
-              subtitle: Text('調整 DPIP 整體的外觀與顏色'.i18n),
+            Selector<SettingsUserInterfaceModel, ThemeMode>(
+              selector: (context, model) => model.themeMode,
+              builder: (context, themeMode, child) {
+                return SegmentedListTile(
+                  isFirst: true,
+                  leading: ContainedIcon(
+                    switch (context.theme.brightness) {
+                      .light => Symbols.light_mode_rounded,
+                      .dark => Symbols.dark_mode_rounded,
+                    },
+                    color: switch (context.theme.brightness) {
+                      .light => Colors.orange,
+                      .dark => Colors.blue[300]!,
+                    },
+                  ),
+                  title: Text('主題模式'.i18n),
+                  subtitle: Text(themeMode.label.i18n),
+                  trailing: const Icon(Symbols.chevron_right_rounded),
+                  onTap: () => const SettingsThemeModeRoute().push(context),
+                );
+              },
             ),
-            const SizedBox(height: 16),
-            SegmentedList(
-              children: [
-                Selector<SettingsUserInterfaceModel, ThemeMode>(
-                  selector: (context, model) => model.themeMode,
-                  builder: (context, themeMode, child) {
-                    return SegmentedListTile(
-                      isFirst: true,
-                      leading: ContainedIcon(
-                        switch (context.theme.brightness) {
-                          .light => Symbols.light_mode_rounded,
-                          .dark => Symbols.dark_mode_rounded,
-                        },
-                        color: switch (context.theme.brightness) {
-                          .light => Colors.orange,
-                          .dark => Colors.blue[300]!,
-                        },
-                      ),
-                      title: Text('主題模式'.i18n),
-                      subtitle: Text(themeMode.label.i18n),
-                      trailing: const Icon(Symbols.chevron_right_rounded),
-                      onTap: () => const SettingsThemeModeRoute().push(context),
-                    );
-                  },
-                ),
-                Selector<SettingsUserInterfaceModel, Color?>(
-                  selector: (context, model) => model.themeColor,
-                  builder: (context, themeColor, child) {
-                    return SegmentedListTile(
-                      isLast: true,
-                      leading: ContainedIcon(
-                        Symbols.colorize_rounded,
-                        color: themeColor ?? context.colors.primary,
-                      ),
-                      title: Text('主題色彩'.i18n),
-                      subtitle: Text(switch (themeColor) {
-                        null => '使用系統配色'.i18n,
-                        final v => ColorTools.nameThatColor(v),
-                      }),
-                      trailing: const Icon(Symbols.chevron_right_rounded),
-                      onTap: () => const SettingsThemeColorRoute().push(context),
-                    );
-                  },
-                ),
-              ],
+            Selector<SettingsUserInterfaceModel, Color?>(
+              selector: (context, model) => model.themeColor,
+              builder: (context, themeColor, child) {
+                return SegmentedListTile(
+                  isLast: true,
+                  leading: ContainedIcon(
+                    Symbols.colorize_rounded,
+                    color: themeColor ?? context.colors.primary,
+                  ),
+                  title: Text('主題色彩'.i18n),
+                  subtitle: Text(switch (themeColor) {
+                    null => '使用系統配色'.i18n,
+                    final v => ColorTools.nameThatColor(v),
+                  }),
+                  trailing: const Icon(Symbols.chevron_right_rounded),
+                  onTap: () => const SettingsThemeColorRoute().push(context),
+                );
+              },
             ),
           ],
-        );
-      },
+        ),
+      ],
     );
   }
 }
