+++ lib/main.dart
@@ -131,6 +131,7 @@ void main() async {
       child: MultiProvider(
         providers: [
           ChangeNotifierProvider.value(value: GlobalProviders.data),
+          ChangeNotifierProvider.value(value: GlobalProviders.experimental),
           ChangeNotifierProvider.value(value: GlobalProviders.location),
           ChangeNotifierProvider.value(value: GlobalProviders.map),
           ChangeNotifierProvider.value(value: GlobalProviders.notification),
